Implement mysky seed provider
NOTE: This is just a reference. The actual seed provider will have to be on a
new page.

I added the handshake and `getRootSeed()` method, and removed the redirect
stuff. Is there anything else that needs to be updated for this work inside an
iframe?

This still needs to be tested.