Consistent use of BigInt in setRegistry
Skynet requires revision IDs to be BigInts. The problem comes when trying to set new registry entries:

```
Expected parameter int to be type bigint, was type number
    at t.assertUint64 (skyid.js:formatted:6724)
    at t.SkynetClient.t.setEntry (skyid.js:formatted:6454)
    at window.SkyID.setRegistry (skyid.js:formatted:10845)
```

This is in the context of the next version of Hacker Paste. Hacker Paste does nothing to manually set revision numbers. And I notice in `skyid.js` that BigInt is not used consistently, or uses implicit conversion. I propose changing the function to consistently, and explicitly, use BigInt.